+++ tests/test_conf.py
@@ -9,3 +9,4 @@ from django_bird.conf import app_settings
 def test_app_settings():
     assert app_settings.COMPONENT_DIRS == []
     assert app_settings.ENABLE_BIRD_ATTRS is True
+    assert app_settings.DEFAULT_ONLY is False